backend/requirements.txt
aiogram==2.23.1
google-api-python-client==2.105.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
loguru==0.7.2
pydantic~=2.4.0
pydantic-settings==2.0.3
Django==4.1.3
django-pandas==0.6.6
pandas~=1.5.1
aiohttp~=3.8.1
uvicorn~=0.23.2
fastapi~=0.103.1